Quell uninitialised-value warning when no "is"
Omitting "is" (rather than using `is => 'bare'`) is useful when extending
an attributed defined in a base class or a role. Previously, the "has"
generated by MooseX::MungeHas was emitting uninitialized-value warnings
when a non-Moo class tried to do that. (Moo classes aren't affected by
this, because Moo natively supports `is => 'lazy'` and `is => 'rwp'`, and
so MooseX::MungeHas doesn't need to look at "is" for Moo.)